requirements.txt
langchain>=0.1.16
transformers>=4.40.0
sentence-transformers>=2.2.2
torch>=2.1.0
pinecone-client>=3.0.0
python-docx>=1.1.0
nltk>=3.8.1
python-dotenv>=1.0.1

# Optional for API serving (Flask or FastAPI, pick one)
flask>=3.0.2
# OR
# fastapi>=0.110.0
# uvicorn[standard]>=0.29.0
